Remove teardown from new phase block (#29)
We are deleting twice a node, before test run and after test run.
While this is fine for sanity, it's inefficient.
However, the real reason for removing this behaviour is because with
the inclusion of platform agnostic modules Nodepool cannot create
the multiple nodes with the current timeout. Increasing the timeout
to build all nodes would make the test runs even longer.